refactor(web-shell): extract shared contexts
