+++ packages/web-shell/client/components/artifacts/SubagentDetail.integration.test.tsx
@@ -75,14 +75,14 @@ vi.mock('../../hooks/useSessionArtifacts', () => ({
   useSessionArtifacts: () => ({ artifacts: [] }),
 }));
 
-vi.mock('../../App', async () => {
+vi.mock('../../WebShellContexts', async () => {
   const { createContext } = await import('react');
   return { CompactModeContext: createContext(false) };
 });
 
 vi.mock('../MessageList', async () => {
   const React = await import('react');
-  const { CompactModeContext } = await import('../../App');
+  const { CompactModeContext } = await import('../../WebShellContexts');
   return {
     MessageList: (props: Record<string, unknown>) => {
       latestMessageListProps.current = props;
@@ -96,7 +96,7 @@ vi.mock('../MessageList', async () => {
 });
 
 const { SubagentDetail } = await import('./SubagentDetail');
-const { CompactModeContext } = await import('../../App');
+const { CompactModeContext } = await import('../../WebShellContexts');
 
 Object.assign(globalThis, { IS_REACT_ACT_ENVIRONMENT: true });
