+++ packages/web-shell/client/components/messages/AssistantMessage.test.tsx
@@ -7,7 +7,7 @@ import { I18nProvider } from '../../i18n';
 
 Object.assign(globalThis, { IS_REACT_ACT_ENVIRONMENT: true });
 
-vi.mock('../../App', async () => {
+vi.mock('../../WebShellContexts', async () => {
   const { createContext } = await import('react');
   return {
     CompactModeContext: createContext(false),
