+++ packages/web-shell/client/components/messages/PlanMessage.tsx
@@ -1,6 +1,6 @@
 import { memo, useContext, useState } from 'react';
 import type { TodoItem } from '../../adapters/types';
-import { TodoTimelineContext } from '../../App';
+import { TodoTimelineContext } from '../../WebShellContexts';
 import { TodoEventSummary, TodoFullList } from './TodoView';
 import { useI18n } from '../../i18n';
 import flashStyles from '../MessageLocateFlash.module.css';
