+++ packages/web-shell/client/components/messages/TodoView.test.tsx
@@ -6,15 +6,14 @@ import { I18nProvider } from '../../i18n';
 import type { TodoItem } from '../../adapters/types';
 import { todoStateKey, type TodoDetail } from '../../utils/todos';
 
-// TodoFullList reads TodoDetailContext from App; mock it so the unit test
-// doesn't pull the whole application graph and can inject its own detail map.
-vi.mock('../../App', async () => {
+// Mock TodoDetailContext so the unit test can inject its own detail map.
+vi.mock('../../WebShellContexts', async () => {
   const { createContext } = await import('react');
   return { TodoDetailContext: createContext(new Map()) };
 });
 
 const { TodoFullList } = await import('./TodoView');
-const { TodoDetailContext } = await import('../../App');
+const { TodoDetailContext } = await import('../../WebShellContexts');
 
 (
   globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }
