+++ packages/web-shell/client/components/messages/tools/SubAgentPanel.test.tsx
@@ -6,10 +6,8 @@ import { I18nProvider } from '../../../i18n';
 import type { ACPToolCall } from '../../../adapters/types';
 import { formatTimestamp } from '../../MessageTimestamp';
 
-// SubAgentPanel pulls in ToolGroup, which imports App for TodoTimelineContext;
-// loading the real App module would drag the whole application graph into this
-// unit test.
-vi.mock('../../../App', async () => {
+// SubAgentPanel pulls in ToolGroup, which reads both todo contexts.
+vi.mock('../../../WebShellContexts', async () => {
   const { createContext } = await import('react');
   return {
     TodoTimelineContext: createContext(new Map()),
